Proxy wasm functionality from content-scripts instead, due to CSP
