+++ apps/extension/src/static/popup.html
@@ -11,7 +11,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <iframe id="sandbox" src="./sandbox.html"></iframe>
   </body>
   <script type="module" src="../popup/index.ts"></script>
 </html>
