--- apps/extension/src/static/sandbox.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="”robots”" content="”noindex,nofollow”" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="description" content="Anoma Extension - Sandbox" />
-    <title>Anoma Extension - Sandbox</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-  </body>
-  <script type="module" src="../sandbox/init.ts"></script>
-</html>
