+++ apps/extension/src/wasm/index.ts
@@ -1,0 +1,1 @@
+export * from "./WasmProxy";
